refactor(fspy-shm): fold shared.rs into mod.rs
mod.rs stopped earning its keep as a separate facade: the wrapper
functions are gone, the curated surface decayed into verbatim
re-exports, and the boundary itself manufactured the cfg(test) re-export
of the error types, needed only so tests one file away could name them.
Merge the mechanism into the module root — overview docs, region views
and ordering contract, writer side, receiver side, then the tests — and
keep the boundary that still means something: layout.rs describes the
region, mod.rs operates on it.
