+++ crates/fspy_shared/src/ipc/channel/shm_io/README.md
@@ -171,19 +171,10 @@ CLAIMED (slot 0) ---+
 
 ## Files
 
-| File        | Role                                                                                                                                                                                                                                        |
-| ----------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `mod.rs`    | Public surface (`ShmWriter`, `ShmReceiver`, `Frames`), the protocol overview docs, and the integration tests — they run against a mocked region and are miri-clean (`cargo miri test -p fspy_shared shm_io`).                               |
-| `shared.rs` | Everything that touches the mapping, in reading order: the typed views of the region and the ordering contract, then the writer side (claim, fill, finish), then the receiver side (close and `Frames`) with the reasoning for its borrows. |
-| `layout.rs` | The region's shape: the header struct, the sizing rule that turns a mapping length into table and payload bounds, payload rounding, span validation, and the descriptor codec. Describes memory, never touches it.                          |
+| File        | Role                                                                                                                                                                                                                                                                         |
+| ----------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `mod.rs`    | The protocol, in reading order: the overview docs, the typed views of the region and the ordering contract, the writer side (claim, fill, finish), the receiver side (close and `Frames`), and the integration tests — miri-clean (`cargo miri test -p fspy_shared shm_io`). |
+| `layout.rs` | The region's shape: the header struct, the sizing rule that turns a mapping length into table and payload bounds, payload rounding, span validation, and the descriptor codec. Describes memory, never touches it.                                                           |
 
-Arrows point at what a file depends on:
-
-```mermaid
-graph TD
-    mod["mod.rs<br>public surface"] --> shared["shared.rs<br>everything touching the mapping"]
-    shared --> layout["layout.rs<br>the region's shape"]
-```
-
-Read from the bottom up — `layout.rs`, then `shared.rs`, then `mod.rs` —
-and each file only needs the ones below it.
+`mod.rs` depends on `layout.rs`; read `layout.rs` first, then `mod.rs` top
+to bottom.
